fixed factorial overflow, now need to fix/add tests

+++ rust/src/physics/single_chain/fjc/thermodynamics/isometric/mod.rs
@@ -59,15 +59,15 @@ impl Isometric for FJC
     fn nondimensional_equilibrium_distribution(&self, nondimensional_end_to_end_length_per_link: &f64) -> f64
     {
         let mut sum: f64 = 0.0;
-        let n = self.number_of_links as u32;
+        let n = self.number_of_links as u128;
         let p = self.number_of_links_f64 - 2.0;
         let m = -*nondimensional_end_to_end_length_per_link*0.5 + 0.5;
-        let k = (self.number_of_links_f64*m).floor() as u32;
+        let k = (self.number_of_links_f64*m).floor() as u128;
         for s in 0..k
         {
             sum += (-1.0_f64).powf(s as f64)*((factorial(n.into())/factorial(s.into())/factorial((n - s).into())) as f64)*(m - (s as f64)/self.number_of_links_f64).powf(p);
         }
-        0.125/PI*nondimensional_end_to_end_length_per_link*(n.pow(n - 2) as f64)/(factorial((n - 2).into()) as f64)*sum
+        0.125/PI*nondimensional_end_to_end_length_per_link*(n.pow((n - 2) as u32) as f64)/(factorial((n - 2).into()) as f64)*sum
     }
     fn equilibrium_radial_distribution(&self, end_to_end_length: &f64) -> f64
     {
@@ -76,15 +76,15 @@ impl Isometric for FJC
     fn nondimensional_equilibrium_radial_distribution(&self, nondimensional_end_to_end_length_per_link: &f64) -> f64
     {
         let mut sum: f64 = 0.0;
-        let n = self.number_of_links as u32;
+        let n = self.number_of_links as u128;
         let p = self.number_of_links_f64 - 2.0;
         let m = -*nondimensional_end_to_end_length_per_link*0.5 + 0.5;
-        let k = (self.number_of_links_f64*m).floor() as u32;
+        let k = (self.number_of_links_f64*m).floor() as u128;
         for s in 0..k
         {
             sum += (-1.0_f64).powf(s as f64)*((factorial(n.into())/factorial(s.into())/factorial((n - s).into())) as f64)*(m - (s as f64)/self.number_of_links_f64).powf(p);
         }
-        0.5*nondimensional_end_to_end_length_per_link*(n.pow(n) as f64)/(factorial((n - 2).into()) as f64)*sum
+        0.5*nondimensional_end_to_end_length_per_link*(n.pow(n as u32) as f64)/(factorial((n - 2).into()) as f64)*sum
     }
 }
